Added omni charm and timeless vault.

+++ src/main/resources/legacy/images/relics/clean_relic_image.py
@@ -1,6 +1,7 @@
 import argparse
 import cv2
 import numpy as np
+import os
 
 def center_image(image, canvas_width, canvas_height):
   # Get image dimensions
@@ -18,19 +19,18 @@ def center_image(image, canvas_width, canvas_height):
 
   return canvas
 
-image = cv2.imread("icon_224.png", cv2.IMREAD_UNCHANGED)
-
-centered_image = center_image(image, 128, 128)
-
-cv2.imwrite("test.png", centered_image)
-
 if __name__ == "__main__":
   parser = argparse.ArgumentParser(description="Make a relic image!")
   parser.add_argument("--input", help="Path to input image", required=True)
   parser.add_argument("--name", help="Name of the relic to write.")
   args = parser.parse_args()
 
+  if not os.path.isfile(args.input):
+    print(f"Can't find file {args.input}")
+    exit()
+
   im = cv2.imread(args.input, cv2.IMREAD_UNCHANGED)
+  im = cv2.resize(im, (64, 64))
   center = center_image(im, 128, 128)
   cv2.imwrite(f"{args.name}.png", center)
